+++ assets/javascripts/kitten/components/cards/crowdfunding-card/test.js
@@ -2,11 +2,12 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
 import { CrowdfundingCard } from './index'
+import { CrowdfundingCardWidget } from './widget'
 
 describe('<CrowdfundingCard />', () => {
   let component
 
-  describe('with some props', () => {
+  describe('CrowdfundingCard props', () => {
     beforeEach(() => {
       component = renderer
         .create(
@@ -25,6 +26,34 @@ describe('<CrowdfundingCard />', () => {
             info2="Lorem ipsum…"
             info3="Lorem ipsum…"
             progress={42}
+          />,
+        )
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('CrowdfundingCardWidget props', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(
+          <CrowdfundingCardWidget
+            ownerTitle="Owner title"
+            ownerDescription="Owner description"
+            imageProps={{ src: '#image', alt: '' }}
+            avatarProps={{ src: '#image', alt: '' }}
+            state="Sur"
+            loading={false}
+            title="Lorem ipsum…"
+            subtitle="Sed ut perspiciatis unde omnis iste natus error Lorem"
+            titleTruncate
+            subTitleTruncate
+            buttonText="Lorem ipsum"
+            counterDay="27days"
+            stateDay=""
           />,
         )
         .toJSON()