+++ assets/javascripts/kitten/components/icons/icons.stories.js
@@ -50,6 +50,7 @@ import { WarningIcon } from './warning-icon'
 import { WarningCircleIcon } from './warning-circle-icon'
 import { WhatsAppIcon } from './whatsapp-icon'
 import { YoutubeIcon } from './youtube-icon'
+import { KissKissBankBankIcon } from './kisskissbankbank-icon'
 
 import { Text } from '../..'
 import React from 'react'
@@ -151,6 +152,7 @@ storiesOf('Icons/List', module).add('default', () => {
         <IconContainer children={<BubbleIcon />} />
         <IconContainer children={<ExportIcon />} />
         <IconContainer children={<StrokeIcon />} />
+        <IconContainer children={<KissKissBankBankIcon width="130" />} />
       </Group>
     </Container>
   )