+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update truncated title on `CrowdfundingCardWidget` component.
 - Fix: Update `Label` styles.
 - Feature: Create `KissKissBankBankLogo` component.
 - Feature: Create `CrowdfundingCardWidget` component.