+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `ToggleSwitch` with styled-components.
+
 ## [2.33.0] - 2019-10-24
 
 Features: